Added User operations and tests
