+++ src/main/java/com/jardessouza/bookstoremanager/author/controller/AuthorController.java
@@ -1,0 +1,45 @@
+package com.jardessouza.bookstoremanager.author.controller;
+
+import com.jardessouza.bookstoremanager.author.dto.AuthorDTO;
+import com.jardessouza.bookstoremanager.author.service.AuthorService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/v1/authors")
+public class AuthorController implements AuthorControllerDocs {
+
+    private AuthorService authorService;
+
+    @Autowired
+    public AuthorController(AuthorService authorService) {
+        this.authorService = authorService;
+    }
+
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public AuthorDTO create(@RequestBody @Valid AuthorDTO authorDTO) {
+        return authorService.create(authorDTO);
+    }
+
+    @GetMapping(path = "/{id}")
+    public AuthorDTO findById(@PathVariable Long id) {
+        return authorService.findById(id);
+    }
+
+    @GetMapping
+    public List<AuthorDTO> listAll() {
+        return this.authorService.findAll();
+    }
+
+    @DeleteMapping(path = "/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void delete(@PathVariable Long id) {
+        this.authorService.delete(id);
+    }
+}
