+++ src/main/java/com/jardessouza/bookstoremanager/author/dto/AuthorDTO.java
@@ -1,0 +1,26 @@
+package com.jardessouza.bookstoremanager.author.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthorDTO {
+
+    private Long id;
+
+    @NotNull
+    @NotEmpty
+    @Size(max = 255)
+    private String name;
+
+    @NotNull
+    @Max(120)
+    private Integer age;
+}
