+++ src/main/java/com/jardessouza/bookstoremanager/author/exception/AuthorNotFoundException.java
@@ -1,0 +1,9 @@
+package com.jardessouza.bookstoremanager.author.exception;
+
+import javax.persistence.EntityNotFoundException;
+
+public class AuthorNotFoundException extends EntityNotFoundException {
+    public AuthorNotFoundException(Long id) {
+        super(String.format("Author with id %s not exists", id));
+    }
+}
