+++ src/main/java/com/jardessouza/bookstoremanager/author/repository/AuthorRepository.java
@@ -1,0 +1,10 @@
+package com.jardessouza.bookstoremanager.author.repository;
+
+import com.jardessouza.bookstoremanager.author.entity.Author;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface AuthorRepository extends JpaRepository<Author, Long> {
+    Optional<Author> findByName(String name);
+}
