+++ src/main/java/com/jardessouza/bookstoremanager/books/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package com.jardessouza.bookstoremanager.books.repository;
+
+import com.jardessouza.bookstoremanager.books.entity.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
