+++ src/main/java/com/jardessouza/bookstoremanager/entity/Auditable.java
@@ -1,0 +1,30 @@
+package com.jardessouza.bookstoremanager.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.Column;
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public abstract class Auditable {
+
+    @CreationTimestamp
+    @Column(name = "createdDate", nullable = false)
+    protected LocalDateTime createdDate;
+    @UpdateTimestamp
+    @Column
+    protected LocalDateTime lastModifiedDate;
+
+}
