--- src/main/java/com/jardessouza/bookstoremanager/model/authorentity/exception/AuthorNotFoundException.java
@@ -1,9 +1,0 @@
-package com.jardessouza.bookstoremanager.model.authorentity.exception;
-
-import javax.persistence.EntityNotFoundException;
-
-public class AuthorNotFoundException extends EntityNotFoundException {
-    public AuthorNotFoundException(Long id) {
-        super(String.format("Author with id %s not exists", id));
-    }
-}
