--- src/main/java/com/jardessouza/bookstoremanager/model/authorentity/repository/AuthorRepository.java
@@ -1,10 +1,0 @@
-package com.jardessouza.bookstoremanager.model.authorentity.repository;
-
-import com.jardessouza.bookstoremanager.model.authorentity.entity.Author;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface AuthorRepository extends JpaRepository<Author, Long> {
-    Optional<Author> findByName(String name);
-}
