--- src/main/java/com/jardessouza/bookstoremanager/model/pusblisherentity/repository/PublisherRepository.java
@@ -1,7 +1,0 @@
-package com.jardessouza.bookstoremanager.model.pusblisherentity.repository;
-
-import com.jardessouza.bookstoremanager.model.pusblisherentity.entity.Publisher;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PublisherRepository extends JpaRepository<Publisher, Long> {
-}
