+++ src/main/java/com/jardessouza/bookstoremanager/pusblisher/dto/PublisherDTO.java
@@ -1,0 +1,30 @@
+package com.jardessouza.bookstoremanager.pusblisher.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+import java.time.LocalDate;
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class PublisherDTO {
+
+    @NotNull
+    private Long id;
+    @NotBlank
+    @Size(max = 205)
+    private String name;
+    @NotBlank
+    @Size(max = 50)
+    private String code;
+    @NotNull
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
+    private LocalDate foundationDate;
+}
