+++ src/main/java/com/jardessouza/bookstoremanager/pusblisher/repository/PublisherRepository.java
@@ -1,0 +1,7 @@
+package com.jardessouza.bookstoremanager.pusblisher.repository;
+
+import com.jardessouza.bookstoremanager.pusblisher.entity.Publisher;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PublisherRepository extends JpaRepository<Publisher, Long> {
+}
