+++ src/main/java/com/jardessouza/bookstoremanager/user/enums/Gender.java
@@ -1,0 +1,16 @@
+package com.jardessouza.bookstoremanager.user.enums;
+
+import lombok.Getter;
+
+
+@Getter
+public enum Gender {
+    MALE("Male"),
+    FEMALE("FEMALE");
+
+    private String description;
+    Gender(String description) {
+        this.description = description;
+    }
+
+}
