+++ src/main/java/com/jardessouza/bookstoremanager/user/exception/UserAlreadyExistsException.java
@@ -1,0 +1,9 @@
+package com.jardessouza.bookstoremanager.user.exception;
+
+import javax.persistence.EntityExistsException;
+
+public class UserAlreadyExistsException extends EntityExistsException {
+    public UserAlreadyExistsException(String name){
+        super(String.format("User with name %s exists", name));
+    }
+}
