+++ src/main/java/com/jardessouza/bookstoremanager/user/exception/UserNotFoundException.java
@@ -1,0 +1,10 @@
+package com.jardessouza.bookstoremanager.user.exception;
+
+import javax.persistence.EntityNotFoundException;
+
+public class UserNotFoundException extends EntityNotFoundException {
+
+    public UserNotFoundException(Long id){
+        super(String.format("User with %d, not exists", id));
+    }
+}
