+++ src/main/java/com/jardessouza/bookstoremanager/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.jardessouza.bookstoremanager.user.repository;
+
+import com.jardessouza.bookstoremanager.user.dto.UserDTO;
+import com.jardessouza.bookstoremanager.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    List<User> findByName(UserDTO userDTO);
+    Optional<User> findByName(String name);
+}
