+++ src/test/java/com/jardessouza/bookstoremanager/publisher/service/PublishServiceTest.java
@@ -1,10 +1,10 @@
 package com.jardessouza.bookstoremanager.publisher.service;
 
-import com.jardessouza.bookstoremanager.model.pusblisherentity.dto.PublisherDTO;
-import com.jardessouza.bookstoremanager.model.pusblisherentity.entity.Publisher;
-import com.jardessouza.bookstoremanager.model.pusblisherentity.mapper.PublisherMapper;
-import com.jardessouza.bookstoremanager.model.pusblisherentity.repository.PublisherRepository;
-import com.jardessouza.bookstoremanager.model.pusblisherentity.service.PublisherService;
+import com.jardessouza.bookstoremanager.pusblisher.dto.PublisherDTO;
+import com.jardessouza.bookstoremanager.pusblisher.entity.Publisher;
+import com.jardessouza.bookstoremanager.pusblisher.mapper.PublisherMapper;
+import com.jardessouza.bookstoremanager.pusblisher.repository.PublisherRepository;
+import com.jardessouza.bookstoremanager.pusblisher.service.PublisherService;
 import com.jardessouza.bookstoremanager.publisher.builder.PublisherDTOBuilder;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
